osd: extract collect_sys_info() out of OSD::_collect_metadata()
