+++ src/common/util.cc
@@ -13,8 +13,11 @@
  */
 
 #include <errno.h>
+#include <sys/utsname.h>
+#include <boost/lexical_cast.hpp>
 
 #include "include/util.h"
+#include "common/debug.h"
 #include "common/errno.h"
 #include "common/strtol.h"
 
@@ -125,4 +128,110 @@ int get_fs_stats(ceph_data_stats_t &stats, const char *path)
   stats.byte_avail = stbuf.f_bavail * stbuf.f_bsize;
   stats.avail_percent = (((float)stats.byte_avail/stats.byte_total)*100);
   return 0;
+}
+
+static bool lsb_release_set(char *buf, const char *prefix,
+			    map<string, string> *pm, const char *key)
+{
+  if (strncmp(buf, prefix, strlen(prefix))) {
+    return false;
+  }
+
+  if (buf[strlen(buf)-1] == '\n')
+    buf[strlen(buf)-1] = '\0';
+
+  char *value = buf + strlen(prefix) + 1;
+  (*pm)[key] = value;
+  return true;
+}
+
+static void lsb_release_parse(map<string, string> *m, CephContext *cct)
+{
+  FILE *fp = popen("lsb_release -idrc", "r");
+  if (!fp) {
+    int ret = -errno;
+    lderr(cct) << "lsb_release_parse - failed to call lsb_release binary with error: " << cpp_strerror(ret) << dendl;
+    return;
+  }
+
+  char buf[512];
+  while (fgets(buf, sizeof(buf) - 1, fp) != NULL) {
+    if (lsb_release_set(buf, "Distributor ID:", m, "distro"))
+      continue;
+    if (lsb_release_set(buf, "Description:", m, "distro_description"))
+      continue;
+    if (lsb_release_set(buf, "Release:", m, "distro_version"))
+      continue;
+    if (lsb_release_set(buf, "Codename:", m, "distro_codename"))
+      continue;
+
+    lderr(cct) << "unhandled output: " << buf << dendl;
+  }
+
+  if (pclose(fp)) {
+    int ret = -errno;
+    lderr(cct) << "lsb_release_parse - pclose failed: " << cpp_strerror(ret) << dendl;
+  }
+}
+
+void collect_sys_info(map<string, string> *m, CephContext *cct)
+{
+  // kernel info
+  struct utsname u;
+  int r = uname(&u);
+  if (r >= 0) {
+    (*m)["os"] = u.sysname;
+    (*m)["kernel_version"] = u.release;
+    (*m)["kernel_description"] = u.version;
+    (*m)["hostname"] = u.nodename;
+    (*m)["arch"] = u.machine;
+  }
+
+  // memory
+  FILE *f = fopen("/proc/meminfo", "r");
+  if (f) {
+    char buf[100];
+    while (!feof(f)) {
+      char *line = fgets(buf, sizeof(buf), f);
+      if (!line)
+	break;
+      char key[40];
+      long long value;
+      int r = sscanf(line, "%s %lld", key, &value);
+      if (r == 2) {
+	if (strcmp(key, "MemTotal:") == 0)
+	  (*m)["mem_total_kb"] = boost::lexical_cast<string>(value);
+	else if (strcmp(key, "SwapTotal:") == 0)
+	  (*m)["mem_swap_kb"] = boost::lexical_cast<string>(value);
+      }
+    }
+    fclose(f);
+  }
+
+  // processor
+  f = fopen("/proc/cpuinfo", "r");
+  if (f) {
+    char buf[100];
+    while (!feof(f)) {
+      char *line = fgets(buf, sizeof(buf), f);
+      if (!line)
+	break;
+      if (strncmp(line, "model name", 10) == 0) {
+	char *c = strchr(buf, ':');
+	c++;
+	while (*c == ' ')
+	  ++c;
+	char *nl = c;
+	while (*nl != '\n')
+	  ++nl;
+	*nl = '\0';
+	(*m)["cpu"] = c;
+	break;
+      }
+    }
+    fclose(f);
+  }
+
+  // distro info
+  lsb_release_parse(m, cct);
 }
